requirements.txt
Flask
Flask-SocketIO
eventlet
pigpio
picamera2
RPi.GPIO
rpi-ws281x
